#29 Move tag dao model to cost package
