+++ lib/infrastructure/local_storage/domain/cost/cost_db_repository.dart
@@ -8,7 +8,7 @@ import 'package:flutter_template/domain/cost/tag.dart';
 import 'package:flutter_template/domain/cost/tags.dart';
 import 'package:flutter_template/domain/cost/title.dart';
 import 'package:flutter_template/infrastructure/local_storage/domain/cost/cost_dao_model.dart';
-import 'package:flutter_template/infrastructure/local_storage/domain/tag/tag_dao_model.dart';
+import 'package:flutter_template/infrastructure/local_storage/domain/cost/tag_dao_model.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
